Add files via upload
All files added via upload. Includes quicksg.deb package for amd64 Ubuntu. Additionally, contains source and compiled c executable contained within the usual /quicksg/usr/bin directory.
You can build your own package by copying the directory structure and building the package with 
$ sudo dpkg-deb --build ~<folder_u_downloaded_to>/quicksg/
and then installing the package by
$ sudo dpkg -i <folder_u_downloaded_to>/quicksg.deb
Note: if you are in the directory a simple ls should tell you if the .deb is present.

Alternatively, if you want to avoid all that building maemae, simply take the quicksg.deb file and install it with dpkg or apt.
Let me know if it works, and if it doesn't.